VEOSS-1430 | Fix completion check issue for FE when part of VAT group. Move EuDetails checks to own file.
