+++ app/utils/CompletionChecks.scala
@@ -21,13 +21,12 @@ import models.euDetails._
 import models.previousRegistrations.PreviousRegistrationDetailsWithOptionalVatNumber
 import models.requests.AuthenticatedDataRequest
 import pages._
-import pages.euDetails._
 import pages.previousRegistrations._
 import play.api.mvc.Results.Redirect
 import play.api.mvc.{AnyContent, Result}
 import queries.previousRegistration._
-import queries.{AllEuOptionalDetailsQuery, AllTradingNames, AllWebsites, EuOptionalDetailsQuery}
-import queries.previousRegistration.AllPreviousRegistrationsWithOptionalVatNumberQuery
+import queries._
+import utils.EuDetailsCompletionChecks._
 
 import scala.concurrent.Future
 
@@ -53,60 +52,6 @@ trait CompletionChecks {
     } else {
       onFailure(incomplete)
     }
-  }
-
-  def getIncompleteEuDetails(index: Index)(implicit request: AuthenticatedDataRequest[AnyContent]): Option[EuOptionalDetails] = {
-    val isPartOfVatGroup = request.userAnswers.vatInfo.exists(_.partOfVatGroup)
-    request.userAnswers
-      .get(EuOptionalDetailsQuery(index))
-      .find(details =>
-        partOfVatGroup(isPartOfVatGroup, details) || notPartOfVatGroup(isPartOfVatGroup, details))
-  }
-
-  def getAllIncompleteEuDetails()(implicit request: AuthenticatedDataRequest[AnyContent]): Seq[EuOptionalDetails] = {
-    val isPartOfVatGroup = request.userAnswers.vatInfo.exists(_.partOfVatGroup)
-    request.userAnswers
-      .get(AllEuOptionalDetailsQuery).map(
-      _.filter(details =>
-        partOfVatGroup(isPartOfVatGroup, details) ||
-        notPartOfVatGroup(isPartOfVatGroup, details)
-      )
-    ).getOrElse(List.empty)
-  }
-
-  private def partOfVatGroup(isPartOfVatGroup: Boolean, details: EuOptionalDetails): Boolean = {
-    isPartOfVatGroup && notSellingToEuConsumers(details) || sellsToEuConsumers(isPartOfVatGroup, details)
-  }
-
-  private def notPartOfVatGroup(isPartOfVatGroup: Boolean, details: EuOptionalDetails): Boolean = {
-    !isPartOfVatGroup && notSellingToEuConsumers(details) || sellsToEuConsumers(isPartOfVatGroup, details)
-  }
-
-  private def notSellingToEuConsumers(details: EuOptionalDetails): Boolean = {
-    details.sellsGoodsToEUConsumers.isEmpty ||
-      (details.sellsGoodsToEUConsumers.contains(false) && details.vatRegistered.isEmpty) ||
-      (details.vatRegistered.contains(true) && details.euVatNumber.isEmpty)
-  }
-
-  private def sellsToEuConsumers(isPartOfVatGroup: Boolean, details: EuOptionalDetails): Boolean = {
-    (details.sellsGoodsToEUConsumers.contains(true) && details.sellsGoodsToEUConsumerMethod.isEmpty) ||
-      (details.sellsGoodsToEUConsumerMethod.contains(EuConsumerSalesMethod.DispatchWarehouse) && details.registrationType.isEmpty) ||
-      (details.registrationType.contains(RegistrationType.VatNumber) && details.euVatNumber.isEmpty) ||
-      (details.registrationType.contains(RegistrationType.TaxId) && details.euTaxReference.isEmpty) ||
-    fixedEstablishment(isPartOfVatGroup, details) || sendsGoods(details)
-
-  }
-
-  private def sendsGoods(details: EuOptionalDetails): Boolean = {
-    (details.sellsGoodsToEUConsumerMethod.contains(EuConsumerSalesMethod.DispatchWarehouse) &&
-      (details.registrationType.contains(RegistrationType.TaxId) || details.registrationType.contains(RegistrationType.VatNumber)) &&
-      (details.euSendGoodsTradingName.isEmpty || details.euSendGoodsAddress.isEmpty))
-  }
-
-  private def fixedEstablishment(isPartOfVatGroup: Boolean, details: EuOptionalDetails): Boolean = {
-    (!isPartOfVatGroup && details.sellsGoodsToEUConsumerMethod.contains(EuConsumerSalesMethod.FixedEstablishment) &&
-      (details.registrationType.contains(RegistrationType.TaxId) || details.registrationType.contains(RegistrationType.VatNumber)) &&
-      (details.fixedEstablishmentTradingName.isEmpty || details.fixedEstablishmentAddress.isEmpty))
   }
 
   def getAllIncompleteDeregisteredDetails()(implicit request: AuthenticatedDataRequest[AnyContent]): Seq[PreviousRegistrationDetailsWithOptionalVatNumber] = {
@@ -153,13 +98,6 @@ trait CompletionChecks {
     }
   }
 
-  def isEuDetailsPopulated()(implicit request: AuthenticatedDataRequest[AnyContent]): Boolean = {
-    request.userAnswers.get(TaxRegisteredInEuPage).exists {
-      case true => request.userAnswers.get(AllEuOptionalDetailsQuery).isDefined
-      case false => true
-    }
-  }
-
   def isDeregisteredPopulated()(implicit request: AuthenticatedDataRequest[AnyContent]): Boolean = {
     request.userAnswers.get(PreviouslyRegisteredPage).exists {
       case true => request.userAnswers.get(AllPreviousRegistrationsWithOptionalVatNumberQuery).isDefined
@@ -186,142 +124,6 @@ trait CompletionChecks {
       incompletePreviousRegistrationRedirect(CheckMode) ++
       incompleteWebsiteUrlsRedirect
       ).headOption
-  }
-
-//  private def incompleteEuDetailsRedirect()(implicit request: AuthenticatedDataRequest[AnyContent]): Option[Result] =
-//    firstIndexedIncompleteEuDetails(getAllIncompleteEuDetails().map(
-//      _.euCountry
-//    )).map(
-//      incompleteCountry =>
-//        Redirect(controllers.euDetails.routes.CheckEuDetailsAnswersController.onPageLoad(CheckMode, Index(incompleteCountry._2)))
-//    )
-
-  //TODO - Change method name when done
-  private def incompleteEuDetailsRedirect(mode: Mode)
-                                            (implicit request: AuthenticatedDataRequest[AnyContent]): Option[Result] = {
-    val isPartOfVatGroup = request.userAnswers.vatInfo.exists(_.partOfVatGroup)
-    firstIndexedIncompleteEuDetails(getAllIncompleteEuDetails().map(_.euCountry)) match {
-      case Some(incompleteCountry) =>
-        request.userAnswers.get(SellsGoodsToEUConsumersPage(Index(incompleteCountry._2))) match {
-          case Some(true) =>
-            request.userAnswers.get(SellsGoodsToEUConsumerMethodPage(Index(incompleteCountry._2))) match {
-              case Some(EuConsumerSalesMethod.FixedEstablishment) =>
-                fixedEstablishmentRedirect(mode, isPartOfVatGroup, incompleteCountry)
-              case Some(EuConsumerSalesMethod.DispatchWarehouse) =>
-                dispatchWarehouseRedirect(mode, incompleteCountry)
-              case None =>
-                Some(Redirect(controllers.euDetails.routes.SellsGoodsToEUConsumerMethodController.onPageLoad(mode, Index(incompleteCountry._2))))
-            }
-          case Some(false) =>
-            notSellingGoodsRedirect(mode, incompleteCountry)
-          case None =>
-            Some(Redirect(controllers.euDetails.routes.SellsGoodsToEUConsumersController.onPageLoad(CheckMode, Index(incompleteCountry._2))))
-
-        }
-    }
-  }
-
-  private def fixedEstablishmentRedirect(
-                                          mode: Mode,
-                                          isPartOfVatGroup: Boolean,
-                                          incompleteCountry: (EuOptionalDetails, Int)
-                                        )(implicit request: AuthenticatedDataRequest[AnyContent]): Option[Result] = {
-    if (isPartOfVatGroup) {
-      // TODO - redirect in controller - create Page instead???
-      ???
-    } else {
-      request.userAnswers.get(RegistrationTypePage(Index(incompleteCountry._2))) match {
-        case Some(RegistrationType.VatNumber) =>
-          request.userAnswers.get(EuVatNumberPage(Index(incompleteCountry._2))) match {
-            case Some(_) =>
-              fixedEstablishingTradeDetailsRedirect(mode, incompleteCountry)
-            case None =>
-              Some(Redirect(controllers.euDetails.routes.EuVatNumberController.onPageLoad(mode, Index(incompleteCountry._2))))
-          }
-        case Some(RegistrationType.TaxId) =>
-          request.userAnswers.get(EuTaxReferencePage(Index(incompleteCountry._2))) match {
-            case Some(_) =>
-              fixedEstablishingTradeDetailsRedirect(mode, incompleteCountry)
-          }
-        case None =>
-          Some(Redirect(controllers.euDetails.routes.RegistrationTypeController.onPageLoad(mode, Index(incompleteCountry._2))))
-      }
-    }
-  }
-
-  private def fixedEstablishingTradeDetailsRedirect(
-                                                     mode: Mode,
-                                                     incompleteCountry: (EuOptionalDetails, Int)
-                                                   )(implicit request: AuthenticatedDataRequest[AnyContent]): Option[Result] = {
-    request.userAnswers.get(FixedEstablishmentTradingNamePage(Index(incompleteCountry._2))) match {
-      case Some(_) =>
-        request.userAnswers.get(FixedEstablishmentAddressPage(Index(incompleteCountry._2))) match {
-          case Some(_) =>
-            Some(Redirect(controllers.euDetails.routes.CheckEuDetailsAnswersController.onPageLoad(mode, Index(incompleteCountry._2))))
-          case None =>
-            Some(Redirect(controllers.euDetails.routes.FixedEstablishmentAddressController.onPageLoad(mode, Index(incompleteCountry._2))))
-        }
-      case None =>
-        Some(Redirect(controllers.euDetails.routes.FixedEstablishmentTradingNameController.onPageLoad(mode, Index(incompleteCountry._2))))
-    }
-  }
-
-  private def dispatchWarehouseRedirect(
-                                         mode: Mode,
-                                         incompleteCountry: (EuOptionalDetails, Int)
-                                       )(implicit request: AuthenticatedDataRequest[AnyContent]): Option[Result] = {
-    request.userAnswers.get(RegistrationTypePage(Index(incompleteCountry._2))) match {
-      case Some(RegistrationType.VatNumber) =>
-        request.userAnswers.get(EuVatNumberPage(Index(incompleteCountry._2))) match {
-          case Some(_) =>
-            dispatchWarehouseTradeDetailsRedirect(mode, incompleteCountry)
-          case None =>
-            Some(Redirect(controllers.euDetails.routes.EuVatNumberController.onPageLoad(mode, Index(incompleteCountry._2))))
-        }
-      case Some(RegistrationType.TaxId) =>
-        request.userAnswers.get(EuTaxReferencePage(Index(incompleteCountry._2))) match {
-          case Some(_) =>
-            dispatchWarehouseTradeDetailsRedirect(mode, incompleteCountry)
-          case None =>
-            Some(Redirect(controllers.euDetails.routes.EuTaxReferenceController.onPageLoad(mode, Index(incompleteCountry._2))))
-        }
-    }
-  }
-
-  private def dispatchWarehouseTradeDetailsRedirect(
-                                                     mode: Mode,
-                                                     incompleteCountry: (EuOptionalDetails, Int)
-                                                   )(implicit request: AuthenticatedDataRequest[AnyContent]): Option[Result] = {
-    request.userAnswers.get(EuSendGoodsTradingNamePage(Index(incompleteCountry._2))) match {
-      case Some(_) =>
-        request.userAnswers.get(EuSendGoodsAddressPage(Index(incompleteCountry._2))) match {
-          case Some(_) =>
-            Some(Redirect(controllers.euDetails.routes.CheckEuDetailsAnswersController.onPageLoad(mode, Index(incompleteCountry._2))))
-          case None =>
-            Some(Redirect(controllers.euDetails.routes.EuSendGoodsAddressController.onPageLoad(mode, Index(incompleteCountry._2))))
-        }
-      case None =>
-        Some(Redirect(controllers.euDetails.routes.EuSendGoodsTradingNameController.onPageLoad(mode, Index(incompleteCountry._2))))
-    }
-  }
-
-  private def notSellingGoodsRedirect(
-                                       mode: Mode,
-                                       incompleteCountry: (EuOptionalDetails, Int)
-                                     )(implicit request: AuthenticatedDataRequest[AnyContent]): Option[Result] = {
-    request.userAnswers.get(VatRegisteredPage(Index(incompleteCountry._2))) match {
-      case Some(true) =>
-        request.userAnswers.get(EuVatNumberPage(Index(incompleteCountry._2))) match {
-          case Some(_) =>
-            Some(Redirect(controllers.euDetails.routes.CheckEuDetailsAnswersController.onPageLoad(mode, Index(incompleteCountry._2))))
-          case None =>
-            Some(Redirect(controllers.euDetails.routes.EuVatNumberController.onPageLoad(mode, Index(incompleteCountry._2))))
-        }
-      case Some(false) =>
-        Some(Redirect(controllers.euDetails.routes.CheckEuDetailsAnswersController.onPageLoad(mode, Index(incompleteCountry._2))))
-      case None =>
-        Some(Redirect(controllers.euDetails.routes.VatRegisteredController.onPageLoad(mode, Index(incompleteCountry._2))))
-    }
   }
 
   def incompletePreviousRegistrationRedirect(mode: Mode)
@@ -373,12 +175,6 @@ trait CompletionChecks {
     None
   }
 
-  private def emptyEuDetailsRedirect()(implicit request: AuthenticatedDataRequest[AnyContent]): Option[Result] = if (!isEuDetailsPopulated) {
-    Some(Redirect(controllers.euDetails.routes.TaxRegisteredInEuController.onPageLoad(CheckMode)))
-  } else {
-    None
-  }
-
   private def emptyDeregisteredRedirect()(implicit request: AuthenticatedDataRequest[AnyContent]): Option[Result] = if (!isDeregisteredPopulated) {
     Some(Redirect(controllers.previousRegistrations.routes.PreviouslyRegisteredController.onPageLoad(CheckMode)))
   } else {
@@ -386,4 +182,5 @@ trait CompletionChecks {
   }
 
 }
+
