+++ conf/app.routes
@@ -41,8 +41,6 @@ GET         /eu-tax/:index                                                      
 POST        /eu-tax/:index                                                      controllers.euDetails.EuCountryController.onSubmit(mode: Mode = NormalMode, index: Index)
 GET         /check-eu-tax/:index                                                controllers.euDetails.EuCountryController.onPageLoad(mode: Mode = CheckMode, index: Index)
 POST        /check-eu-tax/:index                                                controllers.euDetails.EuCountryController.onSubmit(mode: Mode = CheckMode, index: Index)
-GET         /first-check-eu-tax/:index                                          controllers.euDetails.EuCountryController.onPageLoad(mode: Mode = CheckLoopMode, index: Index)
-POST        /first-check-eu-tax/:index                                          controllers.euDetails.EuCountryController.onSubmit(mode: Mode = CheckLoopMode, index: Index)
 
 GET         /eu-vat/:index                                                      controllers.euDetails.VatRegisteredController.onPageLoad(mode: Mode = NormalMode, index: Index)
 POST        /eu-vat/:index                                                      controllers.euDetails.VatRegisteredController.onSubmit(mode: Mode = NormalMode, index: Index)
