+++ test/pages/euDetails/EuCountryPageSpec.scala
@@ -18,7 +18,7 @@ package pages.euDetails
 
 import base.SpecBase
 import controllers.euDetails.{routes => euRoutes}
-import models.{CheckLoopMode, CheckMode, Country, Index, NormalMode, UserAnswers}
+import models.{CheckMode, Country, Index, NormalMode, UserAnswers}
 import org.scalacheck.Arbitrary.arbitrary
 import org.scalatestplus.scalacheck.ScalaCheckPropertyChecks
 import pages.behaviours.PageBehaviours
@@ -41,65 +41,38 @@ class EuCountryPageSpec extends SpecBase with PageBehaviours with ScalaCheckProp
 
       "to Sells Goods to EU Consumer for the same index" in {
 
-        EuCountryPage(countryIndex).navigate(NormalMode, emptyUserAnswers.copy(vatInfo = Some(vatCustomerInfo.copy(partOfVatGroup = true))))
+        EuCountryPage(countryIndex).navigate(NormalMode, emptyUserAnswers)
           .mustEqual(euRoutes.SellsGoodsToEUConsumersController.onPageLoad(NormalMode, countryIndex))
       }
     }
 
     "must navigate in Check mode" - {
 
-      "when Eu Vat Number has not been answered" - {
+      "when Sells Goods to EU Consumers has not been answered" - {
 
-        "to Eu Vat Number for the same index" in {
+        "to Sells Goods to EU Consumers for the same index" in {
 
-          EuCountryPage(countryIndex).navigate(CheckMode, emptyUserAnswers.copy(vatInfo = Some(vatCustomerInfo.copy(partOfVatGroup = true))))
-            .mustEqual(euRoutes.EuVatNumberController.onPageLoad(CheckMode, countryIndex))
+          EuCountryPage(countryIndex).navigate(CheckMode, emptyUserAnswers)
+            .mustEqual(euRoutes.SellsGoodsToEUConsumersController.onPageLoad(CheckMode, countryIndex))
         }
       }
 
-      "when Eu Vat Number has been answered" - {
+      "when Sells Goods to EU Consumers has been answered" - {
 
-        "to wherever Eu Vat Number would navigate to in Check mode" in {
+        "to wherever Sells Goods to EU Consumers would navigate to in Check mode" in {
 
           forAll(arbitrary[UserAnswers]) {
             baseAnswers =>
 
-              val answers = baseAnswers.copy(vatInfo = Some(vatCustomerInfo.copy(partOfVatGroup = true)))
-                .set(EuVatNumberPage(Index(0)), "1234567").success.value
+              val answers = baseAnswers
+                .set(SellsGoodsToEUConsumersPage(Index(0)), true).success.value
 
               EuCountryPage(countryIndex).navigate(CheckMode, answers)
-                .mustEqual(EuVatNumberPage(Index(0)).navigate(CheckMode, answers))
+                .mustEqual(SellsGoodsToEUConsumersPage(Index(0)).navigate(CheckMode, answers))
           }
         }
       }
     }
 
-    "must navigate in Check Loop mode" - {
-
-      "when Eu Vat Number has not been answered" - {
-
-        "to Eu Vat Number for the same index" in {
-
-          EuCountryPage(countryIndex).navigate(CheckLoopMode, emptyUserAnswers.copy(vatInfo = Some(vatCustomerInfo.copy(partOfVatGroup = true))))
-            .mustEqual(euRoutes.EuVatNumberController.onPageLoad(CheckLoopMode, countryIndex))
-        }
-      }
-
-      "when Eu Vat Number has been answered" - {
-
-        "to wherever Eu Vat Number would navigate to in Check Loop mode" in {
-
-          forAll(arbitrary[UserAnswers]) {
-            baseAnswers =>
-
-              val answers = baseAnswers.copy(vatInfo = Some(vatCustomerInfo.copy(partOfVatGroup = true)))
-                .set(EuVatNumberPage(Index(0)), "1234567").success.value
-
-              EuCountryPage(countryIndex).navigate(CheckLoopMode, answers)
-                .mustEqual(EuVatNumberPage(Index(0)).navigate(CheckLoopMode, answers))
-          }
-        }
-      }
-    }
   }
 }
